activate_video() and new_video_from_orphan()

+++ mythvideos/service.py
@@ -1,4 +1,6 @@
 import configparser
+import datetime
+import iso8601
 import os
 import os.path
 import socket
@@ -18,24 +20,91 @@ types of entities.
 """
 class MythVideoService(object):
     def __init__(self):
-        # Read configuration...
-        config_file = os.path.join(settings.BASE_DIR, 'mtv_settings.cfg')
-        self.cfg = configparser.ConfigParser(interpolation=None)
-        config_files_read = self.cfg.read(config_file)
-        if len(config_files_read) == 0:
-            raise Exception('Could not find config file {}'.format(config_file))
+#         # Read configuration...
+#         config_file = os.path.join(settings.BASE_DIR, 'mtv_settings.cfg')
+#         self.cfg = configparser.ConfigParser(interpolation=None)
+#         config_files_read = self.cfg.read(config_file)
+#         if len(config_files_read) == 0:
+#             raise Exception('Could not find config file {}'.format(config_file))
+        self.api = MythApi()
+    
+    
+    def activate_video(self, data={}):
+        """
+        data is a dict, with the following fields:
+            * filespec - mandatory. This field also may include a subdirectory portion.
+                That is, if the file's name is "123.mpg" and it's in the "History" subdirectory
+                under the top-level video directory, then the filespec would be "History/123.mpg"
+                In the above example, if the video top-level directory was "/mnt/store/mythtv/videos"
+                then the full path to this file would be "/mnt/store/mythtv/videos/History/123.mpg"
+                You would put only "History/123.mpg" in this field.
+            * hostname - optional. If none is supplied, then the default value from the config file
+                will be used.
+            * title - optional. If none is supplied, then the filespec will be used, minus the extension
+                and with underlines converted to spaces and path separation characters ("/" or "\") converted
+                to ": " For example, "History/Elizabeth_I_and_Shakespeare.mpg" would be converted to
+                "History: Elizabeth I and Shakespeare"
+            * subtitle - optional. If none is supplied, then the filespec will be used, minus the extension
+                and with underlines converted to spaces and path separation characters ("/" or "\") converted
+                to ": " For example, "History/Elizabeth_I_and_Shakespeare.mpg" would be converted to
+                "History: Elizabeth I and Shakespeare"
+            * length - optional. This is the duration of the video, in minutes. If none is supplied,
+                then this will be set to 0.
+                An exception will be thrown if the value of this field cannot be cast to an integer.
+            * releasedate - optional. If none is supplied, then this will be set to January 1, 1970. If
+                supplied, then this should be a string in ISO8601 format, for example:
+                "2015-11-28T00:00:00Z"
+                An exception will be thrown if this field is not in the valid format.
+            * contenttype - optional. If none is supplied, then this will be set by the database
+                back end to the table's default for this field. If supplied, then this must be one
+                of the values 'MOVIE', 'TELEVISION', 'ADULT', 'MUSICVIDEO', or 'HOMEVIDEO'
+                
+        Returns:
+            MythVideo object, or None if there was a failure adding the file.
+        """
+        filespec = data.get('filespec', None)
+        if filespec is None or filespec == '':
+            raise ValueError("You must supply the filename of the video to activate.")
+        # If no hostname supplied, use default from api...
+        hostname = data.get('hostname', self.api.server_name)
+        # Use MythTV API call to tell MythTV server to add a record to its db for this file...
+        res = self.api.add_to_mythvideo(filespec, hostname) # returns True on success
+        if res: # If API call succeeds, fetch this new db row and fill in rest of metadata...
+            v = MythVideo.objects.filter(filename=filespec,host=hostname).get()
+            if v:
+                v.title = data.get('title', None)
+                if not v.title:
+                    v.title = os.path.splitext(v.filename)[0].replace('_',' ').replace(os.sep,": ")
+                v.subtitle = data.get('subtitle', None)
+                if not v.subtitle:
+                    v.subtitle = os.path.splitext(v.filename)[0].replace('_',' ').replace(os.sep,": ")
+                v.length = int(data.get('length', 0))
+                ct = data.get('contenttype', None)
+                if ct:
+                    v.contenttype = ct
+                datestring = data.get('releasedate', None)
+                if datestring:
+                    v.releasedate = iso8601.parse_date(datestring)
+                else:
+                    v.releasedate = datetime.datetime(1970,1,1)
+                v.year = v.releasedate.year
+                v.save()
+        else: # could not add file or retrieve video
+            v = None
+        return v
+                
+
+            
         
+            
     def new_video_from_orphan(self, orphan, target_host=None, delete_orphan=False):
         """
         1. Figures out where to put file based on orphan's title and subtitle -
             creates appropriate subdirectory under target host's top-level
             video directory, if that subdirectory doesn't already exist.
         2. Copies orphan's file to destination subdirectory
-        3. Executes MythTV API call to add video by filename and host. This
-            results in the addition of a record to MythTV's db table for this video.
-        4. Get a MythVideo object corresponding to the new db record.
-        5. Set metadata for the MythVideo object, then save it.
-        6. If all above was successful and delete_orphan == True, delete the
+        3. Calls activate_video()
+        4. If all above was successful and delete_orphan == True, delete the
             orphan's disk file and the Orphan ORM instance.
         """
         # Step 1: Derive source and target. Create target directory, if needed.
@@ -56,23 +125,21 @@ class MythVideoService(object):
         # Step 2: Copy file
         shutil.copy2(source_filespec, target_directory, follow_symlinks=True)
         # Step 3: Call MythTV API to add video
+        # Construct data dict to pass to activate_video:
         target_filepath = os.path.join(target_subdir, orphan.filename)
-        if not api.add_to_mythvideo(target_filepath, target_host):
-			# api call returns False if problem
-            raise Exception("Could not add video with filename {} and title {}".format(orphan.filename,orphan.title))
-        # Step 4: Get a MythVideo instance for the new entry
-        # This call should return only 1 instance...
-        newvid = MythVideo.objects.filter(hostname=target_host).filter(filename=target_filepath).get()
-        # Step 5: Set metadata and save
-        newvid.title = orphan.title
-        newvid.subtitle = orphan.subtitle
-        newvid.year = orphan.start_date.year
-        newvid.length = orphan.duration
-        newvid.save()
-        # Step 6: Delete original file, if desired
-        if delete_orphan:
-			os.remove(source_filespec)
-			orphan.delete()
+        data = {
+                'filespec': target_filepath,
+                'hostname': target_host,
+                'title': orphan.title,
+                'subtitle': orphan.subtitle,
+                'releasedate': orphan.start_date.isoformat() + 'Z',
+                'length': orphan.duration, # this value is an integer, in minutes
+                }
+        v = self.activate_video(data)
+        # Step 6: Delete orphan, if desired, but only if activation went OK
+        if delete_orphan and v is not None:
+            os.remove(source_filespec)
+            orphan.delete()
